moves our initial data fetch to user-home. should be in routes lololololololol
